Add jwt auth to post_delete

+++ controllers/postController.js
@@ -68,12 +68,32 @@ exports.post_get = (req, res, next) => {
 };
 
 exports.post_delete = (req, res, next) => {
-  Post.findByIdAndDelete(req.params.id, (err) => {
+  passport.authenticate('jwt', { session: false }, (err, user) => {
     if (err) {
       return next(err);
     }
-    res.redirect('/posts');
-  });
+    if (!user) {
+      return res.status(401).json({
+        message: 'Auth Failed: You need to be logged in to update posts.',
+      });
+    }
+    Post.findById(req.params.id).exec((err, results) => {
+      if (err) {
+        return next(err);
+      }
+      if (!results.authorId.equals(user._id)) {
+        return res.status(400).json({
+          message: 'Auth Failed: You can only delete your own posts.',
+        });
+      }
+      Post.findByIdAndDelete(req.params.id, (err) => {
+        if (err) {
+          return next(err);
+        }
+        res.redirect('/posts');
+      });
+    });
+  })(req, res, next);
 };
 
 exports.post_update = [
